Create Ballot_system.sol
Basically, here we have created a smart contact which works on the principle of voting system. This smart contract provides functionality of voting system in which each person have a certain vote value and these voters are only allowed to be registered by the particular head that is chairperson. They can vote to whoever they want in the time range of voting process which can only be done after registration. And finally if we check for the time and it has reached to the done state we declare the results.